+++ packages/blocks-core/js/libs/wordpress/navigation-link/test/functionality.blocks-navigation-link.e2e.cy.js
@@ -38,7 +38,9 @@ describe(
 
 			cy.get('.block-editor-list-view-tree').within(() => {
 				// Open blocks menu
-				cy.get('[aria-label="Add block"]').first().click();
+				cy.get('[aria-label="Add block"]')
+					.first()
+					.click({ force: true });
 			});
 
 			// search for custom link
@@ -49,7 +51,7 @@ describe(
 			// insert
 			cy.get('button.editor-block-list-item-navigation-link')
 				.last()
-				.click();
+				.click({ force: true });
 
 			// enter link value
 			cy.get('input[type="text"]:focus').type('#test{enter}');
@@ -70,6 +72,13 @@ describe(
 					).should('be.visible');
 				}
 			);
+
+			// Icon extension is active
+			cy.getByDataTest('settings-tab').click();
+			cy.getByAriaLabel('Choose Icon…').should('be.visible');
+
+			// switch back to style tab
+			cy.getByDataTest('style-tab').click();
 
 			//
 			// 1. Inner blocks existence