+++ packages/blocks-core/CHANGELOG.md
@@ -7,6 +7,7 @@
 - Added E2E tests to check icon functionality of `Categories` block.
 - Added E2E tests to check icon functionality of `Page List` block.
 - Added E2E tests to check icon functionality of `Home Link` block.
+- Added E2E tests to check icon functionality of `Navigation Link` block.
 
 ## 1.4.0 (2025-05-19)